// src/api/generated/models/consolidation.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { ConsolidationCreatedByUserId } from './consolidationCreatedByUserId';
import type { ConsolidationStatus } from './consolidationStatus';
import type { ConsolidationTransportMode } from './consolidationTransportMode';
import type { ConsolidationCarrier } from './consolidationCarrier';
import type { ConsolidationTrackingNumber } from './consolidationTrackingNumber';
import type { ConsolidationOrigin } from './consolidationOrigin';
import type { ConsolidationDestination } from './consolidationDestination';
import type { ConsolidationNotes } from './consolidationNotes';
import type { ConsolidationPlannedAt } from './consolidationPlannedAt';
import type { ConsolidationInTransitAt } from './consolidationInTransitAt';
import type { ConsolidationCustomsAt } from './consolidationCustomsAt';
import type { ConsolidationDeliveredAt } from './consolidationDeliveredAt';
import type { ConsolidationBuyerCompany } from './consolidationBuyerCompany';
import type { ConsolidationShippingInvoiceEntry } from './consolidationShippingInvoiceEntry';

export interface Consolidation {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  buyerCompanyId: string;
  createdByUserId: ConsolidationCreatedByUserId;
  status: ConsolidationStatus;
  transportMode: ConsolidationTransportMode;
  carrier: ConsolidationCarrier;
  trackingNumber: ConsolidationTrackingNumber;
  origin: ConsolidationOrigin;
  destination: ConsolidationDestination;
  notes: ConsolidationNotes;
  plannedAt: ConsolidationPlannedAt;
  inTransitAt: ConsolidationInTransitAt;
  customsAt: ConsolidationCustomsAt;
  deliveredAt: ConsolidationDeliveredAt;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  updatedAt: string;
  buyerCompany: ConsolidationBuyerCompany;
  shippingInvoices: ConsolidationShippingInvoiceEntry[];
}